pyproject.toml
[project]
name = "promo-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
